window size type fix

+++ include/TE_core.h
@@ -21,7 +21,7 @@ TE_State* TE_create_state(void (*start)(), void (*update)(double), void (*draw)(
 void TE_set_next_state(TE_State *next_state);
 
 typedef struct {
-    float2 size;
+    int2 size;
     uint8_t full_screen;
 }TE_Window_Info;
